pyproject.toml
[tool.poetry]
name = "langgraph-1"
version = "0.1.0"
description = ""
authors = ["edo1z <endo.yuta@logicky.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.59"
langsmith = "^0.2.3"
langchain-anthropic = "^0.3.0"
python-dotenv = "^1.0.1"
tavily-python = "^0.5.0"
langchain-community = "^0.3.12"
langchain-openai = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
